WIP: Use ID mappings features in storage
Provide configuration file and command line options for setting mappings
to use to set default mappings for permissions in the storage library.

Store ID mapping information in pod sandboxes, default to running
sandboxes and containers using the run-time configured mappings instead
of any defaults which might be left over in the storage library, but
default to attaching containers to the sandbox's user namespace and
using its mappings if the container is using the pod's network
namespace.

Mark files that we bind-mount in read-only as also being nodev and
nosuid, so that the remount that runc does to ensure that they're bound
read-only won't fail due to us inadvertently trying to drop those flags.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>